+++ tests/concurrency/sync_types/__init__.py
@@ -1,4 +1,4 @@
-from thread_factory.concurrency.value_types.sync_string import SyncString
+from thread_factory.concurrency.sync_types.sync_string import SyncString
 #from thread_factory.concurrency.value_types.sync_string import SyncString
 #from thread_factory.concurrency.value_types.sync_string import SyncString