+++ tests/concurrency/sync_types/test_sync_bool/test_sync_bool_advanced.py
@@ -22,9 +22,9 @@ import time
 import unittest
 from decimal import Decimal
 
-from thread_factory.concurrency.value_types.sync_bool  import SyncBool
-from thread_factory.concurrency.value_types.sync_int   import SyncInt
-from thread_factory.concurrency.value_types.sync_float import SyncFloat
+from thread_factory.concurrency.sync_types.sync_bool  import SyncBool
+from thread_factory.concurrency.sync_types.sync_int   import SyncInt
+from thread_factory.concurrency.sync_types.sync_float import SyncFloat
 
 
 # ---------------------------------------------------------------------------