+++ tests/concurrency/sync_types/test_sync_string/test_sync_string.py
@@ -3,7 +3,7 @@ import threading
 import copy
 from concurrent.futures import ThreadPoolExecutor
 
-from thread_factory.concurrency.value_types.sync_string import SyncString
+from thread_factory.concurrency.sync_types.sync_string import SyncString
 
 class TestSyncString(unittest.TestCase):